fixes for client and server side component segregation + devops stuff
