+++ next.config.js
@@ -2,7 +2,8 @@
 const nextConfig = {
   images: {
     domains: ['utfs.io']
-  }
+  },
+  output: 'standalone'
 };
 
 module.exports = nextConfig;
